Add files via upload
A simple and interactive web application that displays Pokémon list with the ability to search and filter Pokémons by name or by id, also each Pokémon when clicked displays a detailed information about that Pokémon in a modal. The modal includes Pokémon stats, abilities, height, weight, and a visual representation of their stats using progress bars. The project is built with React-19 Tailwindcss4 and React-icons.